+++ src/fidesctl/cli/options.py
@@ -115,3 +115,73 @@ def output_directory_option(command: Callable) -> Callable:
         help="The output directory for the data map to be exported to.",
     )(command)
     return command
+
+
+def credentials_id_option(command: Callable) -> Callable:
+    "Use credentials defined within fidesctl config."
+    command = click.option(
+        "--credentials-id",
+        type=str,
+        help="Use credentials defined within fidesctl config",
+    )(command)
+    return command
+
+
+def connection_string_option(command: Callable) -> Callable:
+    "Use connection string option to connect to a database"
+    command = click.option(
+        "--connection-string",
+        type=str,
+        help="Use connection string option to connect to a database",
+    )(command)
+    return command
+
+
+def okta_org_url_option(command: Callable) -> Callable:
+    "Use org url option to connect to okta. Requires options --org-url and --token"
+    command = click.option(
+        "--org-url",
+        type=str,
+        help="Use org url option to connect to okta. Requires options --org-url and --token",
+    )(command)
+    return command
+
+
+def okta_token_option(command: Callable) -> Callable:
+    "Use token option to connect to okta. Requires options --org-url and --token"
+    command = click.option(
+        "--token",
+        type=str,
+        help="Use token option to connect to okta. Requires options --org-url and --token",
+    )(command)
+    return command
+
+
+def aws_access_key_id_option(command: Callable) -> Callable:
+    "Use access key id option to connect to aws. Requires options --access_key_id, access_key and --region"
+    command = click.option(
+        "--access_key_id",
+        type=str,
+        help="Use access key id option to connect to aws. Requires options --access_key_id, --secret_access_key and --region",
+    )(command)
+    return command
+
+
+def aws_secret_access_key_option(command: Callable) -> Callable:
+    "Use access key option to connect to aws. Requires options --access_key_id, --secret_access_key and --region"
+    command = click.option(
+        "--secret_access_key",
+        type=str,
+        help="Use access key option to connect to aws. Requires options --access_key_id, --secret_access_key and --region",
+    )(command)
+    return command
+
+
+def aws_region_option(command: Callable) -> Callable:
+    "Use region option to connect to aws. Requires options --access_key_id, access_key and --region"
+    command = click.option(
+        "--region",
+        type=str,
+        help="Use region option to connect to aws. Requires options --access_key_id, --secret_access_key and --region",
+    )(command)
+    return command